Get rid of LibWCRelayMessageData struct
Originally, we used this so that we could avoid allocating memory to copy data
from each message, so that we could just use the data straight from the
message. When I started working with the protocol however, I found out that
this was not as possible as I thought since many of the types of data can't be
used as-is from the message buffer, and need to be copied and made usable
elsewhere. As such, along with the issues this made with memory management
outside of the parser, we now just allocate memory for everything.