Make extract_event_id() indicate when the message is not an event
In Weechat messages, the ID can either indicate that the message itself is an
event, or that the message was sent in response to a previously sent command to
the server. We can tell this difference by assuming that any identifier that is
unknown is actually an identifier we generated and sent with one of our
commands.